rename mod to 'autodrop'
