+++ changelogs/1.5.4.md
@@ -2,3 +2,4 @@
 - fix mod crashing on archive deletion
 - fix mod dropping all items sometimes
 - fix modmenu screen resetting screen offset
+- rename mod to 'autodrop'
