+++ src/main/java/dev/nyon/autodrop/mixins/MinecraftMixin.java
@@ -1,6 +1,6 @@
-package dev.nyon.simpleautodrop.mixins;
+package dev.nyon.autodrop.mixins;
 
-import dev.nyon.simpleautodrop.SimpleAutoDrop;
+import dev.nyon.autodrop.AutoDrop;
 import net.minecraft.client.Minecraft;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -12,6 +12,6 @@ public class MinecraftMixin {
 
     @Inject(method = "tick", at = @At("TAIL"))
     public void onTick(CallbackInfo ci) {
-        SimpleAutoDrop.INSTANCE.tick((Minecraft) (Object) this);
+        AutoDrop.INSTANCE.tick((Minecraft) (Object) this);
     }
 }
