+++ src/main/kotlin/dev/nyon/autodrop/AutoDrop.kt
@@ -1,8 +1,8 @@
-package dev.nyon.simpleautodrop
+package dev.nyon.autodrop
 
 import com.mojang.blaze3d.platform.InputConstants
-import dev.nyon.simpleautodrop.config.*
-import dev.nyon.simpleautodrop.screen.ConfigScreen
+import dev.nyon.autodrop.config.*
+import dev.nyon.autodrop.screen.ConfigScreen
 import kotlinx.coroutines.*
 import net.fabricmc.api.ClientModInitializer
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper
@@ -19,17 +19,17 @@ lateinit var mcDispatcher: CoroutineDispatcher
 lateinit var mcScope: CoroutineScope
 lateinit var minecraft: Minecraft
 
-object SimpleAutoDrop : ClientModInitializer {
+object AutoDrop : ClientModInitializer {
 
     private val toggleKeyBind = KeyBindingHelper.registerKeyBinding(
         KeyMapping(
-            "Toggle AutoDrop", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_J, "SimpleAutoDrop"
+            "Toggle AutoDrop", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_J, "autodrop"
         )
     )
 
     private val menuKeyBind = KeyBindingHelper.registerKeyBinding(
         KeyMapping(
-            "Open GUI", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_O, "SimpleAutoDrop"
+            "Open GUI", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_O, "autodrop"
         )
     )
 
@@ -38,7 +38,7 @@ object SimpleAutoDrop : ClientModInitializer {
             settings.enabled = !settings.enabled
             saveConfig()
             client.gui.setOverlayMessage(
-                Component.literal("SimpleAutoDrop ${if (settings.enabled) "enabled" else "disabled"}").withStyle(
+                Component.literal("autodrop ${if (settings.enabled) "enabled" else "disabled"}").withStyle(
                     Style.EMPTY.withColor(0xF99147)
                 ), false
             )
