+++ src/main/kotlin/dev/nyon/autodrop/config/ModMenuImpl.kt
@@ -1,8 +1,8 @@
-package dev.nyon.simpleautodrop.config
+package dev.nyon.autodrop.config
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory
 import com.terraformersmc.modmenu.api.ModMenuApi
-import dev.nyon.simpleautodrop.screen.ConfigScreen
+import dev.nyon.autodrop.screen.ConfigScreen
 
 @Suppress("unused")
 class ModMenuImpl : ModMenuApi {
