+++ src/main/kotlin/dev/nyon/autodrop/config/models/Config.kt
@@ -1,0 +1,10 @@
+package dev.nyon.autodrop.config.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Config(
+    var enabled: Boolean = true,
+    val archives: MutableList<Archive> = mutableListOf(),
+    val activeArchives: MutableList<String> = mutableListOf()
+)
