+++ src/main/kotlin/dev/nyon/autodrop/screen/AddItemsScreen.kt
@@ -1,10 +1,10 @@
-package dev.nyon.simpleautodrop.screen
+package dev.nyon.autodrop.screen
 
-import dev.nyon.simpleautodrop.config.reloadArchiveProperties
-import dev.nyon.simpleautodrop.config.saveConfig
-import dev.nyon.simpleautodrop.config.settings
-import dev.nyon.simpleautodrop.screen.archiveEntry.ItemIconWidget
-import dev.nyon.simpleautodrop.util.button
+import dev.nyon.autodrop.config.reloadArchiveProperties
+import dev.nyon.autodrop.config.saveConfig
+import dev.nyon.autodrop.config.settings
+import dev.nyon.autodrop.screen.archiveEntry.ItemIconWidget
+import dev.nyon.autodrop.util.button
 import net.minecraft.client.Minecraft
 import net.minecraft.client.gui.GuiGraphics
 import net.minecraft.client.gui.components.ContainerObjectSelectionList
