+++ src/main/kotlin/dev/nyon/autodrop/screen/archive/ArchiveEntry.kt
@@ -1,8 +1,8 @@
-package dev.nyon.simpleautodrop.screen.archive
+package dev.nyon.autodrop.screen.archive
 
-import dev.nyon.simpleautodrop.config.settings
-import dev.nyon.simpleautodrop.minecraft
-import dev.nyon.simpleautodrop.screen.ConfigScreen
+import dev.nyon.autodrop.config.settings
+import dev.nyon.autodrop.minecraft
+import dev.nyon.autodrop.screen.ConfigScreen
 import net.minecraft.client.gui.GuiGraphics
 import net.minecraft.client.gui.components.ObjectSelectionList
 import net.minecraft.network.chat.Component
