--- src/main/kotlin/dev/nyon/simpleautodrop/config/models/Config.kt
@@ -1,8 +1,0 @@
-package dev.nyon.simpleautodrop.config.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-sealed interface Config<P> {
-    fun transformToNew(): P
-}
