+++ src/main/resources/autodrop.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "dev.nyon.simpleautodrop.mixins",
+  "package": "dev.nyon.autodrop.mixins",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "InventoryMixin"
